backend/pyproject.toml
[project]
name = "desafio-igeos"
version = "0.1.0"
description = "API for dashboards about operation costs on the Sistema Interligado Nacional (SIN) organization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.1",
    "cryptography>=44.0.2",
    "fastapi[standard]>=0.115.11",
    "psycopg[binary]>=3.2.6",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.39",
]

[dependency-groups]
scripts = [
    "google-cloud-bigquery>=3.30.0",
    "sqlalchemy>=2.0.39",
]
dev = [
    "aiosqlite>=0.21.0",
]
